[GStreamer][WebRTC] Move network access to NetworkProcess https://bugs.webkit.org/show_bug.cgi?id=244002
Reviewed by NOBODY (OOPS!).

The libnice ICE agent is now confined into the NetworkProcess. Bi-directional communication is
performed with the WebProcess which makes use of a custom GstWebRTCICE implementation.

This patch introduces a build-time dependency on libnice. Before it was a runtime dependency due to
the usage of libnice in the default GstWebRTCICE backend shipped in GStreamer.