[GStreamer][WebRTC] ASSERT when running imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-addTcpIceCandidate.html https://bugs.webkit.org/show_bug.cgi?id=304637
Reviewed by NOBODY (OOPS!).

Keep track of the SDP string without a= prefix in a String. The previous code was assigning the
substring to a StringView, leading to an ASSERT when attempting to remove spurious space characters
because the underlying string wasn't kept around. This is now also done for old GStreamer runtimes.

Driving-by, remove the un-needed GStreamerIceCandidate wrapper and use RTCIceCandidate directly.

* Source/WebCore/Modules/mediastream/gstreamer/GStreamerMediaEndpoint.cpp:
(WebCore::GStreamerMediaEndpoint::addIceCandidate):
* Source/WebCore/Modules/mediastream/gstreamer/GStreamerMediaEndpoint.h:
* Source/WebCore/Modules/mediastream/gstreamer/GStreamerPeerConnectionBackend.cpp:
(WebCore::GStreamerPeerConnectionBackend::doAddIceCandidate):
* Source/WebCore/Modules/mediastream/gstreamer/GStreamerPeerConnectionBackend.h: